[po-22] extract test helper to create user with orga access
