+++ orga/tests/acceptance/authentication-test.js
@@ -2,6 +2,7 @@ import { module, test } from 'qunit';
 import { visit, currentURL, find, fillIn, click } from '@ember/test-helpers';
 import { setupApplicationTest } from 'ember-qunit';
 import { authenticateSession, currentSession } from 'ember-simple-auth/test-support';
+import { createUserWithOrganizationAccess } from '../helpers/test-init';
 
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
 
@@ -9,12 +10,6 @@ module('Acceptance | authentication', function(hooks) {
 
   setupApplicationTest(hooks);
   setupMirage(hooks);
-
-  let user;
-
-  hooks.beforeEach(() => {
-    user = server.create('user', {firstName: 'Titi', lastName: 'Toto', email: 'titi@toto.com'});
-  });
 
   test('it should redirect user to login page if not logged in', async function(assert) {
     // when
@@ -27,18 +22,11 @@ module('Acceptance | authentication', function(hooks) {
 
   test('it should show user name once user is logged in', async function(assert) {
     // given
-    const organization = server.create('organization', {
-      name: 'Le nom de l\'organization'
-    });
-    const organizationAccess = server.create('organization-access', {
-      organizationId: organization.id,
-      userId: user.id
-    });
-    user.organizationAccesses = [organizationAccess];
+    const user = createUserWithOrganizationAccess();
     server.create('campaign');
 
     await visit('/connexion');
-    await fillIn('#login-email', 'titi@toto.com');
+    await fillIn('#login-email', user.email);
     await fillIn('#login-password', 'secret');
 
     // when
@@ -47,23 +35,16 @@ module('Acceptance | authentication', function(hooks) {
     // then
     assert.ok(currentSession(this.application).get('isAuthenticated'), 'The user is authenticated');
 
-    assert.equal(find('.topbar__user-identification').innerText.trim(), "Titi Toto");
+    assert.equal(find('.topbar__user-identification').innerText.trim(), "Harry Cover");
   });
 
   test('it should redirect user to the campaigns list once logged in', async function(assert) {
     // given
-    const organization = server.create('organization', {
-      name: 'Le nom de l\'organization'
-    });
-    const organizationAccess = server.create('organization-access', {
-      organizationId: organization.id,
-      userId: user.id
-    });
-    user.organizationAccesses = [organizationAccess];
+    const user = createUserWithOrganizationAccess();
     server.create('campaign');
 
     await visit('/connexion');
-    await fillIn('#login-email', 'titi@toto.com');
+    await fillIn('#login-email', user.email);
     await fillIn('#login-password', 'secret');
 
     // when
@@ -76,6 +57,8 @@ module('Acceptance | authentication', function(hooks) {
 
   test('it should let user access requested page if user is already authenticated', async function(assert) {
     // given
+    const user = createUserWithOrganizationAccess();
+
     await authenticateSession({
       user_id: user.id,
       access_token: 'access token',
@@ -84,25 +67,16 @@ module('Acceptance | authentication', function(hooks) {
     });
 
     // when
-    await visit('/');
+    await visit('/campagnes/creation');
 
     // then
-    assert.equal(currentURL(), '/');
+    assert.equal(currentURL(), '/campagnes/creation');
     assert.ok(currentSession(this.application).get('isAuthenticated'), 'The user is authenticated');
-
-    assert.equal(find('.topbar__user-identification').innerText.trim(), "Titi Toto");
   });
 
   test('it should display the organization linked to the connected user', async function(assert) {
     // given
-    const organization = server.create('organization', {
-      name: 'Le nom de l\'organization'
-    });
-    const organizationAccess = server.create('organization-access', {
-      organizationId: organization.id,
-      userId: user.id
-    });
-    user.organizationAccesses = [organizationAccess];
+    const user = createUserWithOrganizationAccess();
 
     await authenticateSession({
       user_id: user.id,
@@ -115,7 +89,7 @@ module('Acceptance | authentication', function(hooks) {
     await visit('/');
 
     // then
-    assert.equal(find('.current-organization-panel__name').innerText.trim(), "Le nom de l'organization");
+    assert.equal(find('.current-organization-panel__name').innerText.trim(), "BRO & Evil Associates");
   });
 
 });
