+++ orga/tests/acceptance/campaign-creation-test.js
@@ -2,23 +2,9 @@ import { module, test } from 'qunit';
 import { click, currentURL, fillIn, visit } from '@ember/test-helpers';
 import { setupApplicationTest } from 'ember-qunit';
 import { authenticateSession } from 'ember-simple-auth/test-support';
+import { createUserWithOrganizationAccess } from '../helpers/test-init';
 
 import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
-
-function _createUserWithOneOrganization() {
-  const user = server.create('user');
-
-  const organization = server.create('organization', {
-    name: 'Le nom de l\'organization'
-  });
-  const organizationAccess = server.create('organization-access', {
-    organizationId: organization.id,
-    userId: user.id
-  });
-
-  user.organizationAccesses = [organizationAccess];
-  return user;
-}
 
 module('Acceptance | Campaign Creation', function(hooks) {
 
@@ -35,7 +21,7 @@ module('Acceptance | Campaign Creation', function(hooks) {
 
   test('it should be accessible for an authenticated user', async function(assert) {
     // given
-    const user = _createUserWithOneOrganization();
+    const user = createUserWithOrganizationAccess();
 
     await authenticateSession({
       user_id: user.id,
@@ -51,7 +37,7 @@ module('Acceptance | Campaign Creation', function(hooks) {
 
   test('it should allow creating a campaign', async function(assert) {
     // given
-    const user = _createUserWithOneOrganization();
+    const user = createUserWithOrganizationAccess();
 
     await authenticateSession({
       user_id: user.id,
@@ -80,7 +66,7 @@ module('Acceptance | Campaign Creation', function(hooks) {
         ]
       }, 500);
 
-    const user = _createUserWithOneOrganization();
+    const user = createUserWithOrganizationAccess();
     await authenticateSession({
       user_id: user.id,
     });
